+++ src/main/java/sic/repository/PagoRepository.java
@@ -30,12 +30,12 @@ public interface PagoRepository extends PagingAndSortingRepository<Pago, Long> {
     Pago findTopByEmpresaOrderByNroPagoDesc(Empresa empresa);
     
     @Query("SELECT SUM(p.monto) FROM FacturaVenta fv INNER JOIN fv.pagos p WHERE fv.cliente.id_Cliente = :idCliente AND p.eliminado = false AND p.fecha <= :hasta")
-    Double getSaldoPagosPorCliente(@Param("idCliente") long idCliente, @Param("hasta") Date hasta);
+    BigDecimal getSaldoPagosPorCliente(@Param("idCliente") long idCliente, @Param("hasta") Date hasta);
     
     List<Pago> findByNotaDebitoAndEliminado(Nota nota, boolean eliminado);
     
     @Query("SELECT SUM(p.monto) FROM Factura f INNER JOIN f.pagos p WHERE f.id_Factura = :idFactura AND p.eliminado = false")
-    Double getTotalPagosDeFactura(@Param("idFactura") long idFactura);
+    BigDecimal getTotalPagosDeFactura(@Param("idFactura") long idFactura);
     
     @Query("SELECT SUM(p.monto) FROM Pago p INNER JOIN p.notaDebito n WHERE n.idNota = :idNota AND p.eliminado = false")
     BigDecimal getTotalPagosDeNota(@Param("idNota") long idNota);
